requirements.txt
discord.py==1.6.0
gcloud-aio-storage==6.2.0
google-api-core==1.30.0
google-auth==1.30.2
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-language==2.1.0
httpx==0.16.1
numpy==1.19.5
pantheon==1.3.1
